refactor: vo여서 dto가아닌 도메인에 위치하도록 수정
