+++ src/main/java/domain/AttendanceCount.java
@@ -1,4 +1,4 @@
-package dto;
+package domain;
 
 public record AttendanceCount(int attendance, int late, int absence) {
