+++ src/main/java/domain/CrewDismiss.java
@@ -1,6 +1,5 @@
 package domain;
 
-import dto.AttendanceCount;
 import java.util.Map;
 
 public class CrewDismiss {
